backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sse-starlette>=1.8.2
yt-dlp>=2023.11.16
spotipy>=2.23.0
mutagen>=1.47.0
Pillow>=10.0.0
httpx>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiofiles>=23.2.1
python-multipart>=0.0.6
